tests

+++ src/main/kotlin/org/malachite/estella/people/api/HrPartnerController.kt
@@ -18,9 +18,10 @@ import java.net.URI
 
 @RestController
 @RequestMapping("/api/hrpartners")
-class HrPartnerController(@Autowired private val hrPartnerService: HrPartnerService,
-                          @Autowired private val organizationService: OrganizationService,
-                          @Autowired private val securityService:SecurityService
+class HrPartnerController(
+    @Autowired private val hrPartnerService: HrPartnerService,
+    @Autowired private val organizationService: OrganizationService,
+    @Autowired private val securityService: SecurityService
 ) {
     @CrossOrigin
     @GetMapping
@@ -38,10 +39,12 @@ class HrPartnerController(@Autowired private val hrPartnerService: HrPartnerServ
 
     @CrossOrigin
     @PostMapping("/addHrPartner")
-    fun addHrPartner(@RequestBody hrPartnerRequest: HrPartnerRequest,
-                     @RequestHeader(EStellaHeaders.jwtToken) jwt:String?): ResponseEntity<Message> {
+    fun addHrPartner(
+        @RequestBody hrPartnerRequest: HrPartnerRequest,
+        @RequestHeader(EStellaHeaders.jwtToken) jwt: String?
+    ): ResponseEntity<Message> {
         val organizationUser = securityService.getUserFromJWT(jwt)
-            ?:return OwnResponses.UNAUTH
+            ?: return OwnResponses.UNAUTH
         val organization = organizationService.getOrganizationByUser(organizationUser)
         val saved: HrPartner = hrPartnerService.registerHrPartner(hrPartnerRequest.toHrPartner(organization))
         return ResponseEntity.created(URI("/api/hrpartners/" + saved.id)).body(SuccessMessage)
@@ -49,13 +52,18 @@ class HrPartnerController(@Autowired private val hrPartnerService: HrPartnerServ
 
     @CrossOrigin
     @DeleteMapping("/{hrPartnerId}")
-    fun deleteHrPartner(@PathVariable("hrPartnerId") hrId: Int) =
-        hrPartnerService.deleteHrPartner(hrId).let {
+    fun deleteHrPartner(
+        @RequestHeader(EStellaHeaders.jwtToken) jwt: String?,
+        @PathVariable("hrPartnerId") hrId: Int
+    ): ResponseEntity<Message> {
+        if (!securityService.checkHrRights(jwt, hrId)) return OwnResponses.UNAUTH
+        return hrPartnerService.deleteHrPartner(hrId).let {
             ResponseEntity(SuccessMessage, HttpStatus.OK)
         }
+    }
 }
 
-data class HrPartnerRequest(val mail:String){
-    fun toHrPartner(organization: Organization):HrPartner =
-        HrPartner(null,organization, User(null,"","",mail))
+data class HrPartnerRequest(val mail: String) {
+    fun toHrPartner(organization: Organization): HrPartner =
+        HrPartner(null, organization, User(null, "", "", mail))
 }
